Organizing and Doors Closing Sequence
Organized the files to be easier to find (e.g. putting the different scripts into folders).

Also added a different feedback system for the player when they try to unlock the door once it's closed. It now has a dialog spoken from the player, and subtitle appearing for 3 seconds. This interaction can only happen every 5 seconds, so the player cannot spam it.