Added ending for 1 - 3 Paintings
Added a scene that takes place in the city within the company's building with the player arriving there and reporting back with the disappointing news of not collecting all the paintings.

To get this scene to activate the player has to exit the mansion after collecting 1 - 3 paintings and exiting the gate.