Update main scene
Added dialogue to chief in the ending sequence for him to congratulate the player on successfully destroying the paintings.

Added the scene switcher so the ending screen will load 10 seconds after the player has been congratulated on successfully completing the mission. The player will then be taken to the main menu 10 seconds thereafter.

Added light sources to the lights outside the mansion.

Looped the fire audio so it does not end abruptly.