Interaction to basement and bedroom
Added an interaction in the basement where at two thresholds a werewolf can be seen (and heard) crawling for 3 seconds, before being deactivated. They will see it crawl from one room to another from a small distance.

Added an interaction to one of the bedrooms (the one connected with a bathroom) when the player initially enters. The player can hear (and see) a person walking outside the window for 16 seconds, before no longer being seen.