Subtitles and Library Event
Added Subtitles to the reaction given from Chief when the Mr. / Mrs. Graps paintings are collected.

Added event in the library, which allows the player to enable / disable the light (through the lightswitch). But also includes an initial event with a count sitting at the table saying a dialog. This event only occurs the first time the player interacts with the lightswitch.

Also includes some cleanup, such as removing unnecessary files and small changes to position of objects (such as interface notifications)