Added Blockage in the Intro
Added a white wall with box with narrative reasoning for why the player can't immediately enter the mansion. This also helps the player not get overlapping notifications, audio and subtitles. The white wall is removed when the initial dialog from the chief is completed (78.5 seconds from the launch of the scene).

This change also encourages the player to explore the property before attempting to enter (if they were to reach the door before the narration is completed)