Added Toggle for Flashlight
Added a button to toggle on/off the flashlight with an indication of how to do it, once the player can enter the mansion. The player can however, activate already from the car, if they know which button to press.

Also, some general improvements to the game, such as fixing the position of the worker painting.